+++ packages/cli/scripts/utils/api.ts
@@ -8,7 +8,7 @@ import {
 } from "@aztec/aztec.js";
 import { ContractInstanceWithAddressSchema } from "@aztec/stdlib/contract";
 import { OTCEscrowContract, getEscrowContract, isTestnet } from "@aztec-otc-desk/contracts";
-import type { Order } from "../../../orderflow-service/src/types/api";
+import type { Order } from "../../../api/src/types/api";
 import {
     weth as wethDeployment,
     usdc as usdcDeployment
